requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
requests
python-dotenv
jinja2